requirements.txt
fastapi==0.104.1
uvicorn~=0.29.0
tensorflow==2.15.0
pillow==10.0.0
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.4.2
python-jose==3.3.0
h5py~=3.12.1
pytest~=8.3.4